Most of the ops in there now

+++ beta/init_flow.py
@@ -11,10 +11,10 @@ def get_sinusoid(N, dtype):
         The DFT of the velocity field as a list of three numpy arrays.
     """
 
-    v_x = np.zeros(shape=[N, N, N//2 + 1], dtype=np.complex64)
-    v_y = np.zeros(shape=[N, N, N//2 + 1], dtype=np.complex64)
+    v_x = np.zeros(shape=[N, N, N//2 + 1], dtype=dtype)
+    v_y = np.zeros(shape=[N, N, N//2 + 1], dtype=dtype)
     v_y[1,   0, 0] = +0.5j*N*N*N
     v_y[N-1, 0, 0] = -0.5j*N*N*N
-    v_z = np.zeros(shape=[N, N, N//2 + 1], dtype=np.complex64)
+    v_z = np.zeros(shape=[N, N, N//2 + 1], dtype=dtype)
 
     return [v_x, v_y, v_z]
